pyproject.toml
[tool.poetry]
name = "valentine-user-bot"
version = "0.1.0"
description = ""
authors = ["Amtilan <149930972+Amtilan@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.38.1"
python-dotenv = "^1.0.1"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
